abstract post(s) action creators & types into thier dirs
