+++ src/actions/actionTypes.js
@@ -28,10 +28,6 @@ export const FETCH_POSTS_REQUEST = 'FETCH_POSTS_REQUEST'
 export const FETCH_POSTS_SUCCESS = 'FETCH_POSTS_SUCCESS'
 export const FETCH_POSTS_FAILURE = 'FETCH_POSTS_FAILURE'
 
-export const CREATE_POST_REQUEST = 'CREATE_POST_REQUEST'
-export const CREATE_POST_SUCCESS = 'CREATE_POST_SUCCESS'
-export const CREATE_POST_FAILURE = 'CREATE_POST_FAILURE'
-
 export const RECEIVE_POST_SUCCESS = 'RECEIVE_POST_SUCCESS'
 
 export const POST_LIKE_REQUEST = 'POST_LIKE_REQUEST'
@@ -100,10 +96,6 @@ export const SET_UNREAD_NOTIFICATION_NUM = 'SET_UNREAD_NOTIFICATION_NUM'
 
 export const INCREMENT_UNREAD_NOTIFICATION_NUM =
   'INCREMENT_UNREAD_NOTIFICATION_NUM'
-
-export const DELETE_POST_REQUEST = 'DELETE_POST_REQUEST'
-export const DELETE_POST_SUCCESS = 'DELETE_POST_SUCCESS'
-export const DELETE_POST_FAILURE = 'DELETE_POST_FAILURE'
 
 export const DELETE_REPLY_REQUEST = 'DELETE_REPLY_REQUEST'
 export const DELETE_REPLY_SUCCESS = 'DELETE_REPLY_SUCCESS'
