+++ src/components/posts/store/postTypes.js
@@ -1,0 +1,7 @@
+export const CREATE_POST_REQUEST = 'CREATE_POST_REQUEST'
+export const CREATE_POST_SUCCESS = 'CREATE_POST_SUCCESS'
+export const CREATE_POST_FAILURE = 'CREATE_POST_FAILURE'
+
+export const DELETE_POST_REQUEST = 'DELETE_POST_REQUEST'
+export const DELETE_POST_SUCCESS = 'DELETE_POST_SUCCESS'
+export const DELETE_POST_FAILURE = 'DELETE_POST_FAILURE'
